+++ src/components/ExpandCollapse.vue
@@ -72,6 +72,7 @@ export default {
 </script>
 
 <style lang="sass">
+/* https://raw.githubusercontent.com/patternfly/patternfly-react/master/packages/patternfly-3/patternfly-react/sass/patternfly-react/_expand-collapse.scss */
 @import "~patternfly/dist/sass/patternfly/color-variables";
 
 .expand-collapse-pf {